Propagating changes to canopy classes

+++ src/winds/CanopyElement.h
@@ -113,7 +113,7 @@ public:
   virtual int getCellFlagCanopy() = 0;
   virtual int getCellFlagWake() = 0;
 
-  std::vector<int> canopy_cell2D, canopy_cell3D; /**< map beteen WINDS grid and canopy grid */
+  std::vector<long> canopy_cell2D, canopy_cell3D; /**< map beteen WINDS grid and canopy grid */
   std::map<int, int> canopy_cellMap2D, canopy_cellMap3D; /**< map beteen WINDS grid and canopy grid */
 
   CanopyType _cType;
